Update importTileMap (WIP)
TileMap uses Tiled software (https://www.mapeditor.org/).
Should check for better gid management.

+++ Graficos1_Engine/Engine/src/Tile Map/TileMap.h
@@ -37,8 +37,8 @@ private:
 	unsigned int textureID;
 	unsigned int rendererID;
 
-	float _imageWidth;
-	float _imageHeight;
+	int _imageWidth;
+	int _imageHeight;
 
 	float convertedPosX;
 	float convertedPosY;
